python/Using Php + python send screenshot to server after 1s/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
MouseInfo==0.1.3
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.0.7
requests==2.31.0
urllib3==2.1.0
